fixed XEmbed of external windows (Qt4)
